买卖单撮合引擎实现

+++ model/order.go
@@ -7,7 +7,7 @@ import (
 
 
 type Order struct {
-	ID 				uint `gorm:"primary_key" json:"id"`
+	ID 				uint64 `gorm:"primary_key" json:"id"`
 	CreatedAt 		time.Time `json:"created_at"`
 	UpdatedAt 		time.Time `json:"update_at"`
