+++ model/trade.go
@@ -4,7 +4,7 @@ import "time"
 
 //撮合成交以后的订单成交记录
 type  Trade struct {
-	ID 				uint `gorm:"primary_key" json:"id"`
+	ID 				uint64 `gorm:"primary_key" json:"id"`
 	CreatedAt 		time.Time `json:"created_at"`
 	UpdatedAt 		time.Time `json:"update_at"`
